Move `NetworkIcon` to its own file. Show full chain name for primary networks when test networks are hidden / unavailable
